requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
Flask==2.2.4
Flask-Session==0.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
